Time and length units (#69)
Every configured quantity that describes a rate or an age now means the
same
thing whatever the step duration and whatever the mesh. Four did not.
One of
them was a live defect in the particle transport.

## Drift rates are per unit root time

Pole drift is a random walk, and the spread of a random walk grows with
the
root of the time it takes. A turn of `axis_drift_rate * step_duration`
each
step therefore gave a total wander over a fixed time `T` of
`rate * sqrt(T * step_duration)`. Half the step, half the variance. The
`PoleDriftConfig` doc said the opposite, and that claim was wrong.

Both rates are now per unit root time. The defaults are retuned so that
a
default step behaves as it did: `15.0` and `7.5` become `1.8` and `0.9`,
which
turn an axis 0.213 radians a step against 0.21, and scale a speed by at
most
0.106 against 0.105. `half_axis_turn` gives the per-step angle, so the
test
reads it instead of recovering it from a rotated vector. The scaling is
exact:
`sqrt(1/4)` is `1/2`, so the turn at a quarter of the step is half the
turn.

## Birth and age are model time

`cell_birth` was `Option<i32>` steps. The prior wrote `-hops` into it as
if one
hop were one step. `SeafloorAge` was steps. `cooling_age`,
`maximum_young_age`,
and `abyssal_age_saturation` were integers. All of them changed meaning
when
the step or the mesh changed. The one coincidence that made a hop and a
step
agree was `DEFAULT_STEP_DURATION` itself, which is the time a plate at
maximum
speed takes to cross one cell of the 65,536-cell mesh. That holds at the
defaults and nowhere else.

Birth is now `Option<f32>` model time. A particle born at step `k` gets
`k * step_duration`. `None` is still original continental crust, so
`is_continental` and `CellCrust` do not change. The prior writes
`-hops * hop_duration`, where the hop duration is
`mean_cell_width / (maximum_angular_speed * radius)`. That is what
`DEFAULT_STEP_DURATION` was defined as, so the prior dates crust the
same way
on any mesh and at any step. `ridge_less_age` stays a hop count, because
that
is what it is, and takes the same hop duration.

The prior reads the configured maximum speed, not the fitted plates.
Every
fitted speed is clamped to that maximum, so it is the fastest plate a
world can
hold and not an approximation of one. A world whose fit came out slow no
longer
re-dates its whole ocean, and there is no motionless case to fall back
from.
The maximum was validated finite, non-negative, and above the minimum,
which
let a configured zero through; it is validated positive now, because a
world in
which no plate can move leaves every stage that measures a length
against a
plate speed with nothing to divide by.

`PlateEvolution` carries `elapsed_time`, and `derive_seafloor_age` has
dropped
the `elapsed_steps` argument a caller had to keep in agreement with the
run.
`cooling_age` is 0.56, `maximum_young_age` is 0.14, and
`abyssal_age_saturation` is 0.112: the same forty, ten, and eight
default steps
they were. The oceanic-peak hill strength is now `1 - age /
maximum_young_age`,
the continuous form of the integer ramp it replaced.

## The trench rule reaches as far as a step travels

`subducts` looked for a convergent edge one hop out from the landing
cell.
Measured at the viewer's defaults, the fastest plate moves 1.01 cell
widths a
step at the start, 1.52 once drift has taken it to the band edge, and
1.85 with
a rift opening added. A particle that landed two cells deep found no
edge and
stacked. At the end of a fifteen-step run, 1,245 oceanic particles sat
in cells
another plate owned. The rule intends only transform and divergent
jitter to
survive there.

`TRANSPORT_REACH_HOPS` is now the one fact the empty-cell search, the
gap
radius bound, and the trench rule are stated against. `subducts` asks
whether a
convergent edge lies within that reach with the loser's plate on its far
side,
and it reuses the ring the empty-cell search builds.

Measured before and after: at the viewer's defaults, stranded oceanic
particles
fall from 1,245 to 0 and subducted particles rise from 12,068 to 15,404.
On the
reference fixture they fall from 29 to 0 and rise from 127 to 195. The
rise is
larger than the fall, because a particle that subducts on the step it
arrives
does not stack on any later step. Owner changes fall from 42,722 to
37,860 and
collided cells from 20,896 to 9,608.

## The step is bounded

`maximum_step_duration` returns `TRANSPORT_REACH_HOPS * cell_width` over
the
fastest speed a run can hold, and `evolve_plate_ownership` rejects a
longer
step with `PlateEvolutionError::StepOutrunsReach`. Past that reach a
step is
not a coarser version of the same run: material jumps trenches without
subducting, gaps open that no search can fill, and deformation is raised
at
boundary positions the plates left partway through the step. The
function lives
beside `TRANSPORT_REACH_HOPS` in `transport.rs`, because it is that
reach
restated as a time.

At the viewer's defaults the bound is 0.01513, so 0.014 passes with
eight
percent of the reach to spare. The viewer's step slider takes its top
from the
same function. It stopped at a fixed ten times the default before, which
moved
ten cells a step.

## What this does not fix

A moving world at half the step and twice the count is not the same
world.
Cells resolve twice as often, so which particle wins a contested cell
and when
a gap opens can differ. What is identical is the meaning of every
configured
quantity. A still world is the exact form of that claim, and
`slicing_a_still_run_twice_as_finely_changes_nothing` holds it: crust
birth,
seafloor age, and base elevation come out bit-identical.

Measured at the viewer's defaults, `(0.014, 15)` against `(0.007, 30)`:
continental cells 18,466 against 18,429, land 16,887 against 16,594,
born
particles 4,851 against 4,887, and subducted particles 15,404 against
16,199.
Continental particles are 19,365 at both, exactly, by construction.

`rift_rate * step_duration` is a per-step probability of a Poisson
process and
already scales correctly. `suture_time` and `full_deformation_time` are
already
model time. None of these change.

## Pins

Ownership on the reference fixture moved, from the drift retune and the
trench
reach. Every birth and age pin moved with the type change, and is
re-pinned
through `quantized_fingerprint` over the time field: the 1/1024 grid
resolves a
step of 0.014 and a hop of 0.0138 comfortably. Base elevation and the
oceanic-peak fingerprint moved from their age input. Taking the hop
duration
off the configured maximum speed moved nothing, because the fit clamps
to that
maximum and both worlds hold a plate at it.

Two fixtures are retuned rather than re-pinned. The reference
base-elevation
config scales `cooling_age` to the 512-cell mesh, the way
`reference_evolution_config` already scales the run that feeds it; a hop
there
is eleven times the default mesh's, and without the scaling four fifths
of that
ocean sits flat on the deep floor. Two viewer fixture seeds move to
seeds whose
climate coupling converges on the new elevation field.

The cache needs no version field. Its key is the generator build
identity,
which hashes the crate sources this change touches.

+++ crates/procgen-geology/src/hotspots.rs
@@ -409,7 +409,7 @@ mod tests {
     struct Fixture {
         mesh: SphereMesh,
         plates: PlatePartition,
-        cell_birth: Vec<Option<i32>>,
+        cell_birth: Vec<Option<f32>>,
         kinematics: PlateKinematics,
     }
